Refactor the code that saves, updates and deletes rows from mysql.innodb_table_stats and mysql.innodb_index_stats to prepare it for being able to rename a table (update the corresponding rows).
* Remove the foreign key from innodb_index_stats because it is a source
  of various deadlocks. Also the absence of FK would allow users to
  temporarily store some rows in innodb_index_stats that do not have
  corresponding rows in innodb_table_stats.

* When saving table and its indexes stats - save the table stats and each
  index stat in a separate transaction to avoid deadlocks.

  The previous mode of operation for saving was:

  BEGIN;
  SELECT FOR UPDATE;
  if not found
    INSERT;
  else
    UPDATE;
  COMMIT;

  Which had the flaw that if there was no entry and two concurrent save
  functions were called, then one of them could get a deadlock, like this:

  t1: SELECT FOR UPDATE; (returns not found)
  t2: SELECT FOR UPDATE; (returns not found)
  t1: INSERT; (hangs, waiting on t2)
  t2: INSERT; (fails because of a deadlock)

  Another way of operation would be:

  BEGIN;
  INSERT;
  if duplicate
    UPDATE;
  COMMIT;

  but that could deadlock with a single DELETE of the corresponding entry in
  the table, like this:

  t1: INSERT; (returns duplicate)
  t2: DELETE; (hangs, waiting on t1)
  t1: UPDATE; (fails because of a deadlock)

  Another solution would have been to synchronize dict_stats_save() and
  dict_stats_delete_table_stats() with a simple mutex, but que_eval_sql()
  does not allow any latches to be held during its execution except
  dict_sys->mutex.

  So we change the save of table stats to:

  BEGIN;
  INSERT;
  COMMIT;
  if duplicate
    BEGIN;
    UPDATE;
    COMMIT;

  At worse, a row could get deleted after INSERT has returned a duplicate error
  and before UPDATE executes, but then the UPDATE will be noop because the
  WHERE clause will match nothing. The effect will be as if the delete executed
  after the end of the save operation.

* When deleting stats, delete the table and the index stats in separate
  transactions.

+++ storage/innobase/dict/dict0stats.cc
@@ -1588,94 +1588,92 @@ dict_stats_save_index_stat(
 	ib_uint64_t	stat_value,	/*!< in: value of the stat */
 	ib_uint64_t*	sample_size,	/*!< in: n pages sampled or NULL */
 	const char*	stat_description,/*!< in: description of the stat */
-	trx_t*		trx,		/*!< in/out: transaction to use */
 	ibool		caller_has_dict_sys_mutex)/*!< in: TRUE if the caller
 					owns dict_sys->mutex */
 {
+	trx_t*		trx;
 	pars_info_t*	pinfo;
 	dberr_t		ret;
 
-	pinfo = pars_info_create();
+#define PREPARE_PINFO_FOR_INDEX_SAVE() \
+do { \
+	pinfo = pars_info_create(); \
+	pars_info_add_str_literal(pinfo, "database_name", database_name); \
+	pars_info_add_str_literal(pinfo, "table_name", table_name); \
+	pars_info_add_str_literal(pinfo, "index_name", index->name); \
+	pars_info_add_int4_literal(pinfo, "last_update", last_update); \
+	pars_info_add_str_literal(pinfo, "stat_name", stat_name); \
+	pars_info_add_ull_literal(pinfo, "stat_value", stat_value); \
+	if (sample_size != NULL) { \
+		pars_info_add_ull_literal(pinfo, "sample_size", *sample_size); \
+	} else { \
+		pars_info_add_literal(pinfo, "sample_size", NULL, \
+				      UNIV_SQL_NULL, DATA_FIXBINARY, 0); \
+	} \
+	pars_info_add_str_literal(pinfo, "stat_description", \
+				  stat_description); \
+} while (0);
 
-	pars_info_add_str_literal(pinfo, "database_name", database_name);
-
-	pars_info_add_str_literal(pinfo, "table_name", table_name);
-
-	pars_info_add_str_literal(pinfo, "index_name", index->name);
-
-	pars_info_add_int4_literal(pinfo, "last_update", last_update);
-
-	pars_info_add_str_literal(pinfo, "stat_name", stat_name);
-
-	pars_info_add_ull_literal(pinfo, "stat_value", stat_value);
-
-	if (sample_size != NULL) {
-		pars_info_add_ull_literal(pinfo, "sample_size", *sample_size);
-	} else {
-		pars_info_add_literal(pinfo, "sample_size", NULL,
-				      UNIV_SQL_NULL, DATA_FIXBINARY, 0);
-	}
-
-	pars_info_add_str_literal(pinfo, "stat_description",
-				  stat_description);
-
-	ret = que_eval_sql(pinfo,
-			   "PROCEDURE INDEX_STATS_SAVE () IS\n"
-			   "dummy CHAR;\n"
-			   "BEGIN\n"
-
-			   "SELECT database_name INTO dummy\n"
-			   "FROM \"" INDEX_STATS_NAME "\"\n"
-			   "WHERE\n"
-			   "database_name = :database_name AND\n"
-			   "table_name = :table_name AND\n"
-			   "index_name = :index_name AND\n"
-			   "stat_name = :stat_name\n"
-			   "FOR UPDATE;\n"
-
-			   "IF (SQL % NOTFOUND) THEN\n"
-			   "  INSERT INTO \"" INDEX_STATS_NAME "\"\n"
-			   "  VALUES\n"
-			   "  (\n"
-			   "  :database_name,\n"
-			   "  :table_name,\n"
-			   "  :index_name,\n"
-			   "  :last_update,\n"
-			   "  :stat_name,\n"
-			   "  :stat_value,\n"
-			   "  :sample_size,\n"
-			   "  :stat_description\n"
-			   "  );\n"
-			   "ELSE\n"
-			   "  UPDATE \"" INDEX_STATS_NAME "\" SET\n"
-			   "  last_update = :last_update,\n"
-			   "  stat_value = :stat_value,\n"
-			   "  sample_size = :sample_size,\n"
-			   "  stat_description = :stat_description\n"
-			   "  WHERE\n"
-			   "  database_name = :database_name AND\n"
-			   "  table_name = :table_name AND\n"
-			   "  index_name = :index_name AND\n"
-			   "  stat_name = :stat_name;\n"
-			   "END IF;\n"
-			   "END;",
+	PREPARE_PINFO_FOR_INDEX_SAVE();
+	trx = trx_allocate_for_background();
+	trx_start_if_not_started(trx);
+	ret = que_eval_sql(
+		pinfo,
+		"PROCEDURE INDEX_STATS_SAVE_INSERT () IS\n"
+		"BEGIN\n"
+		"INSERT INTO \"" INDEX_STATS_NAME "\"\n"
+		"VALUES\n"
+		"(\n"
+		":database_name,\n"
+		":table_name,\n"
+		":index_name,\n"
+		":last_update,\n"
+		":stat_name,\n"
+		":stat_value,\n"
+		":sample_size,\n"
+		":stat_description\n"
+		");\n"
+		"END;",
 		!caller_has_dict_sys_mutex, trx);
-
 	/* pinfo is freed by que_eval_sql() */
+	trx_commit_for_mysql(trx);
+	trx_free_for_background(trx);
+
+	if (ret == DB_DUPLICATE_KEY) {
+		PREPARE_PINFO_FOR_INDEX_SAVE();
+		trx = trx_allocate_for_background();
+		trx_start_if_not_started(trx);
+		ret = que_eval_sql(
+			pinfo,
+			"PROCEDURE INDEX_STATS_SAVE_UPDATE () IS\n"
+			"BEGIN\n"
+			"UPDATE \"" INDEX_STATS_NAME "\" SET\n"
+			"last_update = :last_update,\n"
+			"stat_value = :stat_value,\n"
+			"sample_size = :sample_size,\n"
+			"stat_description = :stat_description\n"
+			"WHERE\n"
+			"database_name = :database_name AND\n"
+			"table_name = :table_name AND\n"
+			"index_name = :index_name AND\n"
+			"stat_name = :stat_name;\n"
+			"END;",
+			!caller_has_dict_sys_mutex, trx);
+		/* pinfo is freed by que_eval_sql() */
+		trx_commit_for_mysql(trx);
+		trx_free_for_background(trx);
+	}
 
 	if (ret != DB_SUCCESS) {
 		char	buf_index[MAX_FULL_NAME_LEN];
 		ut_print_timestamp(stderr);
 		fprintf(stderr,
-			" InnoDB: Error while trying to save index "
-			"statistics for table %s.%s, index %s, "
-			"stat name \"%s\": %s\n",
+			" InnoDB: Cannot save index statistics for table "
+			"%s.%s, index %s, stat name \"%s\": %s\n",
 			database_name, table_name,
 			ut_format_name(index->name, FALSE,
 				       buf_index, sizeof(buf_index)),
 			stat_name, ut_strerr(ret));
-
-		trx->error_state = DB_SUCCESS;
 	}
 
 	return(ret);
@@ -1725,79 +1723,73 @@ dict_stats_save(
 	lint */
 	now = (lint) ut_time();
 
+#define PREPARE_PINFO_FOR_TABLE_SAVE() \
+do { \
+	pinfo = pars_info_create(); \
+	pars_info_add_str_literal(pinfo, "database_name", database_name); \
+	pars_info_add_str_literal(pinfo, "table_name", table_name); \
+	pars_info_add_int4_literal(pinfo, "last_update", now); \
+	pars_info_add_ull_literal(pinfo, "n_rows", table->stat_n_rows); \
+	pars_info_add_ull_literal(pinfo, "clustered_index_size", \
+				  table->stat_clustered_index_size); \
+	pars_info_add_ull_literal(pinfo, "sum_of_other_index_sizes", \
+				  table->stat_sum_of_other_index_sizes); \
+} while (0);
+
+	PREPARE_PINFO_FOR_TABLE_SAVE();
 	trx = trx_allocate_for_background();
-
-	/* Use 'read-uncommitted' so that the SELECTs we execute
-	do not get blocked in case some user has locked the rows we
-	are SELECTing */
-
-	trx->isolation_level = TRX_ISO_READ_UNCOMMITTED;
-
 	trx_start_if_not_started(trx);
-
-	pinfo = pars_info_create();
-
-	pars_info_add_str_literal(pinfo, "database_name", database_name);
-
-	pars_info_add_str_literal(pinfo, "table_name", table_name);
-
-	pars_info_add_int4_literal(pinfo, "last_update", now);
-
-	pars_info_add_ull_literal(pinfo, "n_rows", table->stat_n_rows);
-
-	pars_info_add_ull_literal(pinfo, "clustered_index_size",
-				     table->stat_clustered_index_size);
-
-	pars_info_add_ull_literal(pinfo, "sum_of_other_index_sizes",
-				     table->stat_sum_of_other_index_sizes);
-
-	ret = que_eval_sql(pinfo,
-			   "PROCEDURE TABLE_STATS_SAVE () IS\n"
-			   "dummy CHAR;\n"
-			   "BEGIN\n"
-
-			   "SELECT database_name INTO dummy\n"
-			   "FROM \"" TABLE_STATS_NAME "\"\n"
-			   "WHERE\n"
-			   "database_name = :database_name AND\n"
-			   "table_name = :table_name\n"
-			   "FOR UPDATE;\n"
-
-			   "IF (SQL % NOTFOUND) THEN\n"
-			   "  INSERT INTO \"" TABLE_STATS_NAME "\"\n"
-			   "  VALUES\n"
-			   "  (\n"
-			   "  :database_name,\n"
-			   "  :table_name,\n"
-			   "  :last_update,\n"
-			   "  :n_rows,\n"
-			   "  :clustered_index_size,\n"
-			   "  :sum_of_other_index_sizes\n"
-			   "  );\n"
-			   "ELSE\n"
-			   "  UPDATE \"" TABLE_STATS_NAME "\" SET\n"
-			   "  last_update = :last_update,\n"
-			   "  n_rows = :n_rows,\n"
-			   "  clustered_index_size = :clustered_index_size,\n"
-			   "  sum_of_other_index_sizes = "
-			   "    :sum_of_other_index_sizes\n"
-			   "  WHERE\n"
-			   "  database_name = :database_name AND\n"
-			   "  table_name = :table_name;\n"
-			   "END IF;\n"
-			   "END;",
-			   !caller_has_dict_sys_mutex, trx);
-
+	ret = que_eval_sql(
+		pinfo,
+		"PROCEDURE TABLE_STATS_SAVE_INSERT () IS\n"
+		"BEGIN\n"
+		"INSERT INTO \"" TABLE_STATS_NAME "\"\n"
+		"VALUES\n"
+		"(\n"
+		":database_name,\n"
+		":table_name,\n"
+		":last_update,\n"
+		":n_rows,\n"
+		":clustered_index_size,\n"
+		":sum_of_other_index_sizes\n"
+		");\n"
+		"END;",
+		!caller_has_dict_sys_mutex, trx);
 	/* pinfo is freed by que_eval_sql() */
+	trx_commit_for_mysql(trx);
+	trx_free_for_background(trx);
+
+	if (ret == DB_DUPLICATE_KEY) {
+		PREPARE_PINFO_FOR_TABLE_SAVE();
+		trx = trx_allocate_for_background();
+		trx_start_if_not_started(trx);
+		ret = que_eval_sql(
+			pinfo,
+			"PROCEDURE TABLE_STATS_SAVE_UPDATE () IS\n"
+			"BEGIN\n"
+			"UPDATE \"" TABLE_STATS_NAME "\" SET\n"
+			"last_update = :last_update,\n"
+			"n_rows = :n_rows,\n"
+			"clustered_index_size = :clustered_index_size,\n"
+			"sum_of_other_index_sizes = "
+			"  :sum_of_other_index_sizes\n"
+			"WHERE\n"
+			"database_name = :database_name AND\n"
+			"table_name = :table_name;\n"
+			"END;",
+			!caller_has_dict_sys_mutex, trx);
+		/* pinfo is freed by que_eval_sql() */
+		trx_commit_for_mysql(trx);
+		trx_free_for_background(trx);
+	}
 
 	if (ret != DB_SUCCESS) {
 		ut_print_timestamp(stderr);
 		fprintf(stderr,
-			" InnoDB: Error while trying to save table "
-			"statistics for table %s.%s: %s\n",
+			" InnoDB: Cannot save table statistics for table "
+			"%s.%s: %s\n",
 			database_name, table_name, ut_strerr(ret));
-
-		goto end_rollback;
+		return(ret);
 	}
 
 	for (index = dict_table_get_first_index(table);
@@ -1815,10 +1807,9 @@ dict_stats_save(
 						 NULL,
 						 "Number of pages "
 						 "in the index",
-						 trx,
 						 caller_has_dict_sys_mutex);
 		if (ret != DB_SUCCESS) {
-			goto end_rollback;
+			return(ret);
 		}
 
 		ret = dict_stats_save_index_stat(database_name, table_name,
@@ -1827,10 +1818,9 @@ dict_stats_save(
 						 NULL,
 						 "Number of leaf pages "
 						 "in the index",
-						 trx,
 						 caller_has_dict_sys_mutex);
 		if (ret != DB_SUCCESS) {
-			goto end_rollback;
+			return(ret);
 		}
 
 		n_uniq = dict_index_get_n_unique(index);
@@ -1873,31 +1863,16 @@ dict_stats_save(
 				index, now, stat_name,
 				stat_n_diff_key_vals[i],
 				&stat_n_sample_sizes[i],
-				stat_description, trx,
+				stat_description,
 				caller_has_dict_sys_mutex);
 
 			if (ret != DB_SUCCESS) {
-				goto end_rollback;
+				return(ret);
 			}
 		}
 	}
 
-	trx_commit_for_mysql(trx);
-	ret = DB_SUCCESS;
-	goto end_free;
-
-end_rollback:
-
-	trx->op_info = "rollback of internal transaction on stats tables";
-	trx_rollback_to_savepoint(trx, NULL);
-	trx->op_info = "";
-	ut_a(trx->error_state == DB_SUCCESS);
-
-end_free:
-
-	trx_free_for_background(trx);
-
-	return(ret);
+	return(DB_SUCCESS);
 }
 /* @} */
 
@@ -2830,26 +2805,13 @@ dict_stats_delete_table_stats(
 		return(DB_SUCCESS);
 	}
 
-	/* Create a new private trx */
-
-	trx = trx_allocate_for_background();
-
-	/* Use 'read-uncommitted' so that the SELECTs we execute
-	do not get blocked in case some user has locked the rows we
-	are SELECTing */
-
-	trx->isolation_level = TRX_ISO_READ_UNCOMMITTED;
-
-	trx_start_if_not_started(trx);
-
 	/* Increment table reference count to prevent the tables from
 	being DROPped just before que_eval_sql(). */
 	dict_stats = dict_stats_open();
 
 	if (dict_stats == NULL) {
 		/* stats tables do not exist or have unexpected structure */
-		ret = DB_SUCCESS;
-		goto commit_and_return;
+		return(DB_SUCCESS);
 	}
 
 	ut_snprintf(database_name, sizeof(database_name), "%.*s",
@@ -2859,48 +2821,49 @@ dict_stats_delete_table_stats(
 	table_name_strip = dict_remove_db_name(table_name);
 
 	pinfo = pars_info_create();
-
 	pars_info_add_str_literal(pinfo, "database_name", database_name);
-
 	pars_info_add_str_literal(pinfo, "table_name", table_name_strip);
 
-	ret = que_eval_sql(pinfo,
-			   "PROCEDURE DROP_TABLE_STATS () IS\n"
-			   "dummy CHAR;\n"
-			   "BEGIN\n"
-
-			   /* Lock TABLE_STATS_NAME first to avoid deadlock
-			   with dict_stats_save() which locks the two tables
-			   in order TABLE_STATS_NAME,INDEX_STATS_NAME. The
-			   foreign key imposes that order during INSERT/UPDATE
-			   and the reverse order during DELETE. */
-
-			   "SELECT database_name INTO dummy\n"
-			   "FROM \"" TABLE_STATS_NAME "\"\n"
-			   "WHERE\n"
-			   "database_name = :database_name AND\n"
-			   "table_name = :table_name\n"
-			   "FOR UPDATE;\n"
-
-			   "DELETE FROM \"" INDEX_STATS_NAME "\" WHERE\n"
-			   "database_name = :database_name AND\n"
-			   "table_name = :table_name;\n"
-
-			   "DELETE FROM \"" TABLE_STATS_NAME "\" WHERE\n"
-			   "database_name = :database_name AND\n"
-			   "table_name = :table_name;\n"
-
-			   "END;\n",
-			   TRUE,
-			   trx);
-
+	trx = trx_allocate_for_background();
+	trx_start_if_not_started(trx);
+	ret = que_eval_sql(
+		pinfo,
+		"PROCEDURE DROP_TABLE_STATS_TABLE () IS\n"
+		"BEGIN\n"
+		"DELETE FROM \"" TABLE_STATS_NAME "\" WHERE\n"
+		"database_name = :database_name AND\n"
+		"table_name = :table_name;\n"
+		"END;\n",
+		TRUE, trx);
 	/* pinfo is freed by que_eval_sql() */
+	trx_commit_for_mysql(trx);
+	trx_free_for_background(trx);
+
+	if (ret == DB_SUCCESS) {
+		pinfo = pars_info_create();
+		pars_info_add_str_literal(pinfo, "database_name", database_name);
+		pars_info_add_str_literal(pinfo, "table_name", table_name_strip);
+
+		trx = trx_allocate_for_background();
+		trx_start_if_not_started(trx);
+		ret = que_eval_sql(
+			pinfo,
+			"PROCEDURE DROP_TABLE_STATS_INDEX () IS\n"
+			"BEGIN\n"
+			"DELETE FROM \"" INDEX_STATS_NAME "\" WHERE\n"
+			"database_name = :database_name AND\n"
+			"table_name = :table_name;\n"
+			"END;\n",
+			TRUE, trx);
+		/* pinfo is freed by que_eval_sql() */
+		trx_commit_for_mysql(trx);
+		trx_free_for_background(trx);
+	}
 
 	if (ret != DB_SUCCESS) {
 
 		ut_snprintf(errstr, errstr_sz,
-			    "Unable to delete statistics for table %s.%s "
-			    "from %s or %s%s: %s. "
+			    "Unable to delete statistics for table %s.%s: %s. "
 			    "They can be deleted later using "
 
 			    "DELETE FROM %s WHERE "
@@ -2912,12 +2875,6 @@ dict_stats_delete_table_stats(
 			    "table_name = '%s';",
 
 			    database_name, table_name_strip,
-			    TABLE_STATS_NAME_PRINT, INDEX_STATS_NAME_PRINT,
-
-			    (ret == DB_LOCK_WAIT_TIMEOUT
-			     ? " because the rows are locked"
-			     : ""),
-
 			    ut_strerr(ret),
 
 			    INDEX_STATS_NAME_PRINT,
@@ -2931,12 +2888,6 @@ dict_stats_delete_table_stats(
 	}
 
 	dict_stats_close(dict_stats);
-
-commit_and_return:
-
-	trx_commit_for_mysql(trx);
-
-	trx_free_for_background(trx);
 
 	return(ret);
 }
